2.1

+++ src/main/java/exopandora/worldhandler/builder/component/impl/ComponentAttributeItem.java
@@ -28,7 +28,7 @@ public class ComponentAttributeItem extends ComponentAttribute
 	{
 		NBTTagList attributes = new NBTTagList();
 		
-		for(Entry<EnumAttributes, Float> entry : this.attributes.entrySet())
+		for(Entry<EnumAttributes, Double> entry : this.attributes.entrySet())
 		{
 			if(this.applyable.apply(entry.getKey()) && entry.getValue() != 0)
 			{
